server.js 모듈화
