+++ index.js
@@ -1,0 +1,5 @@
+const {createServer} = require('./src/server.js');
+const {connectSocket} = require('./src/socket.js');
+console.log(createServer)
+const server = createServer();
+connectSocket(server);
