+++ src/server.js
@@ -1,0 +1,21 @@
+const express = require('express');
+const app = express();
+const server = require("http").createServer(app);
+
+const PORT = 3000;
+
+module.exports = {
+  createServer: () => {
+    app.use(express.static("public"));
+
+    app.get('/', (req, res) => {
+      res.sendFile(__dirname + 'public/index.html')
+    })
+    
+    server.listen(PORT, () => {
+      console.log(`listening ${PORT} port`);
+    })
+    
+    return server;
+  }
+}
